requirements.txt
asgiref==3.3.4
Django==3.2.3
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-js-asset==1.2.2
django-mptt==0.12.0
pytz==2021.1
sqlparse==0.4.1
